+++ apps/frontend/src/app/core/components/chat/chat.component.html
@@ -1,0 +1,37 @@
+<div class="flex-1 p-6 overflow-y-auto space-y-4">
+  <div *ngFor="let message of messages" class="flex" [ngClass]="{'justify-end': message.sender === 'user', 'justify-start': message.sender === 'ai'}">
+    <div class="max-w-2xl p-3 rounded-lg shadow-md" [ngClass]="{'bg-blue-500 text-white': message.sender === 'user', 'bg-white text-gray-800': message.sender === 'ai'}">
+      <p *ngIf="!message.isLoading" class="whitespace-pre-wrap">{{ message.text }}</p>
+      <div *ngIf="message.isLoading" class="flex items-center space-x-2">
+          <div class="w-2 h-2 bg-gray-400 rounded-full animate-pulse"></div>
+          <div class="w-2 h-2 bg-gray-400 rounded-full animate-pulse [animation-delay:0.2s]"></div>
+          <div class="w-2 h-2 bg-gray-400 rounded-full animate-pulse [animation-delay:0.4s]"></div>
+      </div>
+    </div>
+  </div>
+</div>
+<div class="p-4 bg-white border-t border-gray-200">
+  <form [formGroup]="chatForm" (ngSubmit)="sendMessage()" class="flex items-center space-x-4">
+    <select formControlName="provider" class="p-2 border rounded-md bg-gray-50 focus:outline-none focus:ring-2 focus:ring-blue-500">
+      <option value="dummy">Dummy</option>
+      <option value="lm-studio">LM Studio (Local)</option>
+      <option value="openai">OpenAI</option>
+      <option value="mistral">Mistral</option>
+      <option value="gemini">Gemini</option>
+      <option value="perplexity">Perplexity</option>
+      <option value="grok">Grok</option>
+      <option value="deepseek">Deepseek</option>
+    </select>
+
+    <input formControlName="apiKey" *ngIf="requiresApiKey(chatForm.get('provider')?.value)" type="password" placeholder="API Key..." class="p-2 border rounded-md flex-shrink w-1/4 focus:outline-none focus:ring-2 focus:ring-blue-500">
+
+    <input formControlName="prompt" type="text" placeholder="Type your message..." class="flex-1 p-2 border rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500">
+
+    <button type="submit"
+    [disabled]="chatForm.invalid || isLoading"
+    class="px-4 py-2 bg-blue-600 text-white font-semibold rounded-md hover:bg-blue-700 disabled:bg-blue-300 disabled:cursor-not-allowed transition-colors">
+      Send
+    </button>
+  </form>
+</div>
+
