+++ apps/frontend/src/app/core/components/layout/layout.component.html
@@ -1,0 +1,27 @@
+<div class="flex h-screen bg-gray-200 font-sans">
+  <aside class="w-64 bg-gray-800 text-white flex flex-col justify-between">
+    <div>
+      <div class="p-4 border-b border-gray-700">
+        <h1 class="text-xl font-bold">Chateroo</h1>
+      </div>
+      <nav class="flex-1 p-2 space-y-2">
+        <a routerLink="/chat" routerLinkActive="bg-gray-700 text-white" [routerLinkActiveOptions]="{exact: true}" class="flex items-center px-4 py-2 rounded-md text-gray-400 hover:bg-gray-700 hover:text-white">
+          <span>Chat</span>
+        </a>
+        <a routerLink="/canvas" routerLinkActive="bg-gray-700 text-white" class="flex items-center px-4 py-2 rounded-md text-gray-400 hover:bg-gray-700 hover:text-white">
+          <span>Canvas</span>
+        </a>
+      </nav>
+    </div>
+    <nav class="p-2 border-t border-gray-700">
+      <a routerLink="/settings" routerLinkActive="bg-gray-700 text-white" class="flex items-center px-4 py-2 rounded-md text-gray-400 hover:bg-gray-700 hover:text-white">
+        <span>Settings</span>
+      </a>
+    </nav>
+  </aside>
+
+  <main class="flex-1 flex flex-col bg-gray-100">
+    <router-outlet></router-outlet>
+  </main>
+</div>
+
