+++ apps/frontend/src/app/core/components/settings/settings.component.html
@@ -1,0 +1,17 @@
+<div class="p-8">
+  <h2 class="text-2xl font-bold text-gray-800 mb-6">Settings</h2>
+  <form [formGroup]="settingsForm" (ngSubmit)="saveSettings()" class="space-y-6 bg-white p-6 rounded-lg shadow-md max-w-lg">
+    <div>
+      <label for="openai-key" class="block text-sm font-medium text-gray-700">
+        OpenAI API Key
+      </label>
+      <input id="openai-key" formControlName="openAiApiKey" type="password" class="mt-1 block w-full p-2 border border-gray-300 rounded-md shadow-sm focus:ring-blue-500 focus:border-blue-500">
+    </div>
+    <div class="flex justify-end">
+      <button type="submit" class="px-4 py-2 bg-blue-600 text-white font-semibold rounded-md hover:bg-blue-700">
+        Save Settings
+      </button>
+    </div>
+  </form>
+</div>
+
